backend/services/transcription/requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-whisper==20250625
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.14.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
